+++ tests/testthat/test-cal-validate.R
@@ -11,6 +11,19 @@ test_that("Logistic validation works", {
 
 })
 
+test_that("Linear validation works", {
+  df <- testthat_cal_reg_sampled()
+
+  res_sum <- cal_validate_linear(df, outcome)
+  expect_s3_class(res_sum, "data.frame")
+  expect_equal(nrow(res_sum), 2)
+
+  res_not_sum <- cal_validate_linear(df, outcome, summarize = FALSE)
+  expect_s3_class(res_not_sum, "data.frame")
+  expect_equal(nrow(res_not_sum), 10)
+
+})
+
 test_that("Isotonic validation works", {
   df <- testthat_cal_sampled()
 
@@ -60,5 +73,25 @@ test_that("Setting summarize to FALSE returns new columns", {
     )
 
   expect_snapshot(no_res)
+
+})
+
+test_that("check metric direction", {
+  library(yardstick)
+  reg_stats <- metric_set(rmse, ccc)
+
+  df <- testthat_cal_reg_sampled()
+
+  res_sum <-
+    cal_validate_linear(df, outcome, metrics = reg_stats) %>%
+    dplyr::distinct(.metric, direction) %>%
+    dplyr::arrange(.metric, direction) %>%
+    dplyr::select(.metric, direction)
+  metric_df <- tibble::as_tibble(reg_stats)%>%
+    dplyr::distinct(.metric = metric, direction) %>%
+    dplyr::arrange(.metric, direction) %>%
+    dplyr::select(.metric, direction)
+
+  expect_equal(res_sum, metric_df)
 
 })